+++ zcash_client_sqlite/src/wallet/init/migrations.rs
@@ -160,6 +160,7 @@ migration_modules!(
     support_zcashd_wallet_import,
     transparent_gap_limit_handling,
     transparent_spend_locator_map,
+    transparent_tx_address_observations,
     tree_retained_checkpoints,
     tx_observation_height,
     tx_retrieval_queue,
@@ -405,6 +406,9 @@ pub(super) fn all_migrations<
         Box::new(v_migration_transactions::Migration),
         Box::new(standalone_address::Migration),
         Box::new(fix_v_transactions_multi_account_totals::Migration),
+        Box::new(transparent_tx_address_observations::Migration {
+            _params: params.clone(),
+        }),
     ]
 }